Add protocol for AccountService
